Update AgentTestRunner to use JUnit5
