+++ dd-java-agent/instrumentation-testing/src/main/groovy/datadog/trace/agent/test/injection/LogContextInjectionTestBase.groovy
@@ -1,4 +1,4 @@
-package datadog.trace.agent.test.log.injection
+package datadog.trace.agent.test.injection
 
 import datadog.trace.agent.test.AgentTestRunner
 import datadog.trace.api.CorrelationIdentifier
