+++ dd-java-agent/instrumentation-testing/src/main/resources/META-INF/services/org.junit.platform.launcher.LauncherSessionListener
@@ -1,0 +1,1 @@
+datadog.trace.agent.test.BootstrapClasspathSetup
