+++ dd-java-agent/instrumentation/build.gradle
@@ -87,35 +87,18 @@ subprojects { Project subProj ->
       //testImplementation project(':dd-java-agent:instrumentation:http-url-connection')
       testImplementation project(':dd-java-agent:instrumentation:classloading')
 
-      testImplementation project(':dd-java-agent:testing')
+      testImplementation project(':dd-java-agent:instrumentation-testing')
       testAnnotationProcessor libs.autoservice.processor
       testCompileOnly libs.autoservice.annotation
 
       instrumentPluginClasspath project(path: ':dd-java-agent:agent-tooling', configuration: 'instrumentPluginClasspath')
     }
 
-    subProj.testing {
-      suites.configureEach {
-        // SpockRunner that we use to run agent tests cannot be properly ported to JUnit 5,
-        // since the framework does not provide the hooks / extension points
-        // that can be used to shadow the tested class.
-
-        // In order to mitigate this, SpockRunner extends JUnitPlatform,
-        // which is a JUnit 4 runner that allows executing JUnit 5 tests in a JUnit 4 environment
-        // (i.e. running them as JUnit 4 tests).
-
-        // So even though Spock 2 tests run on top of JUnit 5,
-        // we execute them in "compatibility mode" so that SpockRunner could shadow the test class
-        // See https://junit.org/junit5/docs/current/user-guide/#running-tests-junit-platform-runner for more details.
-        useJUnit()
-      }
-    }
-
     subProj.tasks.withType(Test).configureEach { subTask ->
       // The `forkedTest` task is not a proper test suite, so it has to be configured directly
       if (subTask.name == 'forkedTest') {
         // See SpockRunner comment above
-        useJUnit()
+        useJUnitPlatform()
       }
 
       if (subTask.name in ['latestDepTest', 'latestDepForkedTest']) {
