+++ dd-java-agent/instrumentation/jetty-11/build.gradle
@@ -43,11 +43,11 @@ dependencies {
 
 
 
-  testFixturesImplementation(project(':dd-java-agent:testing')) {
+  testFixturesImplementation(project(':dd-java-agent:instrumentation-testing')) {
     exclude group: 'org.eclipse.jetty', module: 'jetty-server'
   }
   // Don't want to conflict with jetty from the test server.
-  testImplementation(project(':dd-java-agent:testing')) {
+  testImplementation(project(':dd-java-agent:instrumentation-testing')) {
     exclude group: 'org.eclipse.jetty', module: 'jetty-server'
   }
   testImplementation ("org.eclipse.jetty:jetty-server:11.0.0") {
