+++ dd-java-agent/instrumentation/jetty-12/build.gradle
@@ -63,7 +63,7 @@ dependencies {
   implementation project(':dd-java-agent:instrumentation:jetty-common')
 
   // Don't want to conflict with jetty from the test server.
-  testImplementation(project(':dd-java-agent:testing')) {
+  testImplementation(project(':dd-java-agent:instrumentation-testing')) {
     exclude group: 'org.eclipse.jetty', module: 'jetty-server'
   }
   testImplementation ("org.eclipse.jetty:jetty-server:12.0.0") {
