+++ dd-java-agent/instrumentation/jetty-client/jetty-client-12.0/build.gradle
@@ -31,7 +31,7 @@ dependencies {
   testImplementation(project(':dd-java-agent:instrumentation:jetty-client:jetty-client-9.1'))
   testImplementation(project(':dd-java-agent:instrumentation:jetty-client:jetty-client-10.0'))
 
-  testImplementation(project(path:':dd-java-agent:testing', configuration:'shadow')) {
+  testImplementation(project(path:':dd-java-agent:instrumentation-testing', configuration:'shadow')) {
     // explicitly declared below.
     exclude group: 'org.eclipse.jetty'
   }
