+++ dd-java-agent/instrumentation/kotlin-coroutines/build.gradle
@@ -47,7 +47,7 @@ dependencies {
   testFixturesImplementation libs.coroutines
   testFixturesApi project(':dd-trace-api')
   testFixturesApi project(':dd-java-agent:instrumentation:trace-annotation')
-  testFixturesApi project(':dd-java-agent:testing')
+  testFixturesApi project(':dd-java-agent:instrumentation-testing')
   testFixturesApi 'com.github.spotbugs:spotbugs-annotations:4.2.0'
 
   testImplementation libs.kotlin
