+++ dd-java-agent/instrumentation/resteasy-appsec/build.gradle
@@ -32,11 +32,11 @@ dependencies {
   testFixturesApi group: 'org.jboss.resteasy', name: 'resteasy-jaxrs', version: '3.0.0.Final'
   testFixturesApi group: 'org.jboss.resteasy', name: 'resteasy-jackson-provider', version: '3.0.0.Final'
   testFixturesApi group: 'org.jboss.resteasy', name: 'resteasy-multipart-provider', version: '3.0.0.Final'
-  testFixturesImplementation project(':dd-java-agent:testing'), {
+  testFixturesImplementation project(':dd-java-agent:instrumentation-testing'), {
     exclude group: 'org.eclipse.jetty', module: 'jetty-server'
   }
 
-  testImplementation project(':dd-java-agent:testing'), {
+  testImplementation project(':dd-java-agent:instrumentation-testing'), {
     exclude group: 'org.eclipse.jetty', module: 'jetty-server'
   }
