+++ dd-java-agent/instrumentation/spring-webmvc-3.1/build.gradle
@@ -42,7 +42,7 @@ dependencies {
   //  compileOnly group: 'org.springframework', name: 'spring-webmvc', version: '2.5.6'
   //  compileOnly group: 'javax.servlet', name: 'servlet-api', version: '2.4'
 
-  testImplementation(project(':dd-java-agent:testing')) {
+  testImplementation(project(':dd-java-agent:instrumentation-testing')) {
     exclude(module: 'jetty-server') // incompatible servlet api
   }
